PERA-1381 :: Move some files into transaction.sdk
