+++ common-sdk/src/main/kotlin/com/algorand/wallet/algosdk/transaction/sdk/AlgoTransactionSigner.kt
@@ -10,7 +10,7 @@
  * limitations under the License
  */
 
-package com.algorand.wallet.algosdk.transaction
+package com.algorand.wallet.algosdk.transaction.sdk
 
 interface AlgoTransactionSigner {
