+++ common-sdk/src/main/kotlin/com/algorand/wallet/algosdk/transaction/sdk/SignHdKeyTransaction.kt
